feat: add localstorage for pizza and cart, split selectors, slice and types in redux logic
