+++ src/components/Header.tsx
@@ -1,18 +1,28 @@
+import { useEffect, useRef } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 
 import logoSvg from '../assets/img/pizza-logo.svg';
 import Search from './Search';
-import { selectCart } from '../redux/slices/cartSlice';
+import { selectCart } from '../redux/cart/selectors';
 
 const Header = () => {
     const { items, totalPrice } = useSelector(selectCart);
     const location = useLocation();
+    const isMounted = useRef(false);
 
     const totalCount = items.reduce(
         (sum: number, item: any) => sum + item.count,
         0
     );
+
+    useEffect(() => {
+        if (isMounted) {
+            const json = JSON.stringify(items);
+            localStorage.setItem('cart', json);
+        }
+        isMounted.current = true;
+    }, [items]);
 
     return (
         <div className="header">
