+++ src/components/PizzaBlock/index.tsx
@@ -1,11 +1,10 @@
 import { FC, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import {
-    CartItem,
-    addItem,
-    selectCartItemById,
-} from '../../redux/slices/cartSlice';
 import { Link } from 'react-router-dom';
+
+import { selectCartItemById } from '../../redux/cart/selectors';
+import { CartItem } from '../../redux/cart/types';
+import { addItem } from '../../redux/cart/slice';
 interface IPizzaBlock {
     id: string;
     imageUrl: string;
