+++ src/components/Sort.tsx
@@ -1,6 +1,7 @@
 import { FC, memo, useEffect, useRef, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { Sort, SortPropertyEnum, setSort } from '../redux/slices/filterSlice';
+import { Sort, SortPropertyEnum } from '../redux/filter/types';
+import { setSort } from '../redux/filter/slice';
 
 type SortItem = {
     name: string;
