+++ src/redux/cart/slice.ts
@@ -1,24 +1,9 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { RootState } from '../store';
+import { getCartFromLS } from '../../utils/getCartFromLS';
+import { calcTotalPrice } from '../../utils/calcTotalPrice';
+import { CartItem, CartState } from './types';
 
-export type CartItem = {
-    id: string;
-    title: string;
-    price: number;
-    imageUrl: string;
-    type: string;
-    size: number;
-    count: number;
-};
-export interface CartState {
-    totalPrice: number;
-    items: CartItem[];
-}
-
-const initialState: CartState = {
-    totalPrice: 0,
-    items: [],
-};
+const initialState: CartState = getCartFromLS();
 
 export const cartSlice = createSlice({
     name: 'filter',
@@ -38,25 +23,23 @@ export const cartSlice = createSlice({
                 });
             }
 
-            state.totalPrice = state.items.reduce(
-                (sum: number, obj: any) => obj.price * obj.count + sum,
-                0
-            );
+            state.totalPrice = calcTotalPrice(state.items);
         },
 
         minusItem(state, action: PayloadAction<string>) {
             const findItem = state.items.find(
-                (obj: any) => obj.id === action.payload
+                (obj) => obj.id === action.payload
             );
 
             if (findItem) {
                 findItem.count--;
             }
+            state.totalPrice = calcTotalPrice(state.items);
         },
 
         removeItem(state, action: PayloadAction<string>) {
             state.items = state.items.filter(
-                (obj: any) => obj.id !== action.payload
+                (obj) => obj.id !== action.payload
             );
         },
         clearItems(state) {
@@ -65,10 +48,6 @@ export const cartSlice = createSlice({
         },
     },
 });
-
-export const selectCart = (state: RootState) => state.cart;
-export const selectCartItemById = (id: string) => (state: RootState) =>
-    state.cart.items.find((obj) => obj.id === id);
 
 export const { addItem, minusItem, removeItem, clearItems } = cartSlice.actions;
