+++ src/redux/cart/types.ts
@@ -1,0 +1,13 @@
+export type CartItem = {
+    id: string;
+    title: string;
+    price: number;
+    imageUrl: string;
+    type: string;
+    size: number;
+    count: number;
+};
+export interface CartState {
+    totalPrice: number;
+    items: CartItem[];
+}
