+++ src/redux/filter/slice.ts
@@ -1,31 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { RootState } from '../store';
-
-export enum SortPropertyEnum {
-    RATING_DESC = 'rating',
-    RATING_ASC = '-rating',
-    TITLE_DESC = 'title',
-    TITLE_ASC = '-title',
-    PRICE_DESC = 'price',
-    PRICE_ASC = '-price',
-}
-
-export type Sort = {
-    name: string;
-    sortProperty:
-        | 'rating'
-        | 'title'
-        | 'price'
-        | '-rating'
-        | '-title'
-        | '-price';
-};
-export interface FilterState {
-    searchValue: string;
-    categoryId: number;
-    currentPage: number;
-    sort: Sort;
-}
+import { FilterState, Sort, SortPropertyEnum } from './types';
 
 const initialState: FilterState = {
     searchValue: '',
@@ -69,9 +43,6 @@ export const filterSlice = createSlice({
         },
     },
 });
-
-export const selectFilter = (state: RootState) => state.filter;
-export const selectSort = (state: RootState) => state.filter.sort;
 
 export const {
     setCategoryId,
