--- src/redux/slices/pizzaSlice.ts
@@ -1,71 +1,0 @@
-import { createSlice, createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
-import axios from 'axios';
-import { RootState } from '../store';
-
-type Pizza = {
-    id: string;
-    imageUrl: string;
-    title: string;
-    price: number;
-    sizes: number[];
-    types: number[];
-    rating: number;
-};
-
-export enum Status {
-    LOADING = 'loading',
-    SUCCESS = 'success',
-    ERROR = 'error',
-}
-export interface PizzaState {
-    items: Pizza[];
-    status: Status;
-}
-
-const initialState: PizzaState = {
-    items: [],
-    status: Status.LOADING,
-};
-
-export const fetchPizzas = createAsyncThunk<Pizza[], Record<string, string>>(
-    'pizza/fetchPizzasStatus',
-    async (params) => {
-        const { sortBy, order, category, search, currentPage } = params;
-        const { data } = await axios.get<Pizza[]>(
-            `https://6614489d2fc47b4cf27c0ab5.mockapi.io/items?page=${currentPage}&limit=4&${category}&sortBy=${sortBy}&order=${order}${search}`
-        );
-
-        return data;
-    }
-);
-
-export const pizzaSlice = createSlice({
-    name: 'pizza',
-    initialState,
-    reducers: {
-        setItems(state, action: PayloadAction<Pizza[]>) {
-            state.items = action.payload;
-        },
-    },
-    extraReducers: (builder) => {
-        builder
-            .addCase(fetchPizzas.pending, (state) => {
-                state.status = Status.LOADING;
-                state.items = [];
-            })
-            .addCase(fetchPizzas.fulfilled, (state, action) => {
-                state.items = action.payload;
-                state.status = Status.SUCCESS;
-            })
-            .addCase(fetchPizzas.rejected, (state) => {
-                state.status = Status.ERROR;
-                state.items = [];
-            });
-    },
-});
-
-export const selectPizzaData = (state: RootState) => state.pizza;
-
-export const { setItems } = pizzaSlice.actions;
-
-export default pizzaSlice.reducer;
