+++ src/scss/App.scss
@@ -23,6 +23,22 @@ body {
 .content {
     padding: 40px 0;
 
+    &-error-info {
+        margin: 100px auto;
+        width: 500px;
+        text-align: center;
+
+        h2 {
+            font-size: 32px;
+            font-weight: 700;
+            margin-bottom: 10px;
+        }
+
+        p {
+            font-size: 20px;
+        }
+    }
+
     &-top {
         display: flex;
         align-items: center;
